+++ src/app/components/views/upload-file/upload-file.component.html
@@ -1,5 +1,5 @@
 <input type="file" (change)="uploadFile($event)" />
 
-<div>{{ uploadPercent | async }}</div>
+<!-- <div>{{ uploadPercent | async }}</div> -->
 
 <a [href]="downloadURL | async">{{ downloadURL | async }}</a>
