+++ src/app/components/views/upload-file/upload-file.component.ts
@@ -1,9 +1,8 @@
 // https://github.com/angular/angularfire/blob/master/docs/storage/storage.md
 
 import { Component, OnInit } from '@angular/core';
-import { AngularFireStorage } from '@angular/fire/storage';
-import { Observable } from 'rxjs';
-import { finalize, tap } from 'rxjs/operators';
+// import { Observable } from 'rxjs';
+import { FileUploadService } from '../../../services/file-upload.service'
 
 @Component({
   selector: 'app-upload-file',
@@ -12,27 +11,18 @@ import { finalize, tap } from 'rxjs/operators';
 })
 export class UploadFileComponent implements OnInit {
 
-  uploadPercent: Observable<number>;
-  downloadURL: Observable<string>;
+  // uploadPercent: Observable<number>;
+  // downloadURL: Observable<string>;
+  downloadURL: string;
 
   ngOnInit() {
   }
 
-  constructor(private storage: AngularFireStorage) {}
+  constructor(private m: FileUploadService) {
+  }
 
   uploadFile(event) {
-    const file = event.target.files[0];
-    const filePath = `images/${Date.now()}_${file.name}`;
-    const fileRef = this.storage.ref(filePath);
-    const task = this.storage.upload(filePath, file);
-
-    this.uploadPercent = task.percentageChanges();
-
-    task
-    .snapshotChanges()
-    .pipe(
-      finalize(() => this.downloadURL = fileRef.getDownloadURL())
-    )
-    .subscribe()
+    this.m.uploadFile(event).subscribe(url => { this.downloadURL = url });
   }
+
 }
