+++ src/app/services/file-upload.service.ts
@@ -1,0 +1,37 @@
+// https://github.com/angular/angularfire/blob/master/docs/storage/storage.md
+
+import { Injectable, OnInit } from '@angular/core';
+import { AngularFireStorage } from '@angular/fire/storage';
+import { Observable } from 'rxjs';
+import { finalize, tap } from 'rxjs/operators';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class FileUploadService {
+
+  uploadPercent: Observable<number>;
+  downloadURL: Observable<string>;
+
+  constructor(private storage: AngularFireStorage) {}
+
+  uploadFile(event): Observable<string> {
+    const file = event.target.files[0];
+    const filePath = `images/${Date.now()}_${file.name}`;
+    const fileRef = this.storage.ref(filePath);
+    const task = this.storage.upload(filePath, file);
+
+    this.uploadPercent = task.percentageChanges();
+
+    task
+    .snapshotChanges()
+    .pipe(
+      finalize(() => {
+        this.downloadURL = fileRef.getDownloadURL();
+      })
+    )
+    .subscribe()
+
+    return this.downloadURL;
+  }
+}
